# back/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Canadian Solar ERP API",
    "description": "API complète pour la gestion de magasin solaire et électronique",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "decimal"
          },
          "stockQuantity": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "out_of_stock",
              "on_order"
            ]
          },
          "barcode": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "supplierId": {
            "type": "integer"
          },
          "photo": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Lister tous les produits",
        "responses": {
          "200": {
            "description": "Liste des produits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Créer un produit",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produit créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Récupérer un produit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Produit non trouvé"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Mettre à jour un produit",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit mis à jour"
          },
          "404": {
            "description": "Produit non trouvé"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Supprimer un produit",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Produit supprimé"
          },
          "404": {
            "description": "Produit non trouvé"
          }
        }
      }
    },
...
    "/api/products/{id}/movements": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Historique des mouvements de stock",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mouvements de stock"
          }
        }
      }
    },
    "/api/products/{id}/restock": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Réapprovisionner un produit",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produit réapprovisionné"
          }
        }
      }
    },
    "/api/products/{id}/adjust-stock": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Ajuster manuellement le stock d'un produit",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock ajusté"
          }
        }
      }
    },
...
    "/api/reviews/product/{productId}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Avis d'un produit",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avis du produit"
          }
        }
      }
    },
...
    "/api/movements/product/{productId}": {
      "get": {
        "tags": [
          "Movements"
        ],
        "summary": "Mouvements d’un produit",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mouvements du produit"
          }
        }
      }
    },
...
  }
}